+++ src/emulator/controller.ts
@@ -340,14 +340,6 @@ export async function startAll(options: any, noGui: boolean = false): Promise<vo
 
     const firestoreEmulator = new FirestoreEmulator(args);
     await startEmulator(firestoreEmulator);
-
-    firestoreLogger.logLabeled(
-      "BULLET",
-      "firestore",
-      `For testing set ${clc.bold(
-        `${Constants.FIRESTORE_EMULATOR_HOST}=${firestoreAddr.host}:${firestoreAddr.port}`
-      )}`
-    );
   }
 
   if (shouldStart(options, Emulators.DATABASE)) {
@@ -389,14 +381,6 @@ export async function startAll(options: any, noGui: boolean = false): Promise<vo
 
     const databaseEmulator = new DatabaseEmulator(args);
     await startEmulator(databaseEmulator);
-
-    databaseLogger.logLabeled(
-      "BULLET",
-      "database",
-      `For testing set ${clc.bold(
-        `${Constants.FIREBASE_DATABASE_EMULATOR_HOST}=${databaseAddr.host}:${databaseAddr.port}`
-      )}`
-    );
   }
 
   if (shouldStart(options, Emulators.HOSTING)) {