+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adds new API commands that handle list/create/delete operations on the Android SHA certificate hashes `apps:android:sha:list`, `apps:android:sha:create`, and `apps:android:sha:delete`.
 - Fixes an issue where the CLI did not assume admin privileges when performing Firestore / RTDB emulator operations.
 - Fixes an issue where the functions and hosting emulators would crash when not properly initialized (#2112).
+- Improves logs when running `emulators:start`.
 - Fixes an issue where `use` would allow an invalid project identifier.
 - Fixes an issue where custom options passed to `admin.initializeApp()` in the functions emulator were improperly augmented.
 - Changes `firebasemods.*.*` IAM permission checks to `firebaseextensions.*.*`